+++ pytorch_lightning/trainer/trainer.py
@@ -32,7 +32,7 @@ from pytorch_lightning.trainer.training_loop import TrainerTrainLoopMixin
 from pytorch_lightning.trainer.training_tricks import TrainerTrainingTricksMixin
 from pytorch_lightning.trainer.lr_finder import TrainerLRFinderMixin
 from pytorch_lightning.utilities.exceptions import MisconfigurationException
-from pytorch_lightning.utilities import rank_zero_warn, parsing, rank_zero_info
+from pytorch_lightning.utilities import rank_zero_warn, parsing, rank_zero_info, rank_zero_only
 
 try:
     from apex import amp
@@ -321,6 +321,14 @@ class Trainer(
             # fixing non-deterministic part of horovod
             # https://github.com/PyTorchLightning/pytorch-lightning/pull/1572/files#r420279383
             os.environ["HOROVOD_FUSION_THRESHOLD"] = str(0)
+
+        # init the default rank if exists
+        # we need to call this here or NVIDIA flags and other messaging in init will show on all ranks
+        # this way we only show it on rank 0
+        if 'LOCAL_RANK' in os.environ:
+            rank_zero_only.rank = os.environ['LOCAL_RANK']
+        if 'SLURM_JOB_ID' in os.environ:
+            rank_zero_only.rank = os.environ['SLURM_JOB_ID']
 
         # Init callbacks
         self.prepare_data_per_node = prepare_data_per_node
@@ -843,6 +851,10 @@ class Trainer(
         if self.is_function_implemented('on_fit_start'):
             model.on_fit_start()
 
+        self.setup('fit')
+        if self.is_function_implemented('setup'):
+            model.setup('fit')
+
         # on multi-gpu jobs we only want to manipulate (download, etc) on node_rank=0, local_rank=0
         # or in the case where each node needs to do its own manipulation in which case just local_rank=0
         if self.can_prepare_data():
@@ -888,6 +900,7 @@ class Trainer(
                 mp.spawn(self.ddp_train, nprocs=self.num_processes, args=(model,))
 
             elif self.distributed_backend == 'ddp_spawn':
+                self.__set_random_port()
                 model.share_memory()
 
                 # spin up peers
@@ -944,6 +957,10 @@ class Trainer(
         # model hooks
         if self.is_function_implemented('on_fit_end'):
             model.on_fit_end()
+
+        self.teardown('fit')
+        if self.is_function_implemented('teardown'):
+            model.teardown('fit')
 
         # return 1 when finished
         # used for testing or when we need to know that training succeeded
@@ -1128,6 +1145,11 @@ class Trainer(
             trainer = Trainer()
             trainer.test(model, test_dataloaders=test)
         """
+        self.setup('test')
+        if self.is_function_implemented('setup'):
+            model_ref = self.model if model is None else model
+            model_ref.setup('test')
+
         if model is None and ckpt_path == 'best' and self.checkpoint_callback.save_top_k <= 0:
             raise MisconfigurationException(
                 'ckpt_path is "best", but ModelCheckpoint is not configured to save the best model.')
@@ -1166,6 +1188,10 @@ class Trainer(
             self.run_evaluation(test_mode=True)
 
         self.testing = False
+
+        self.teardown('test')
+        if self.is_function_implemented('teardown'):
+            self.model.teardown('test')
 
     def check_model_configuration(self, model: LightningModule):
         r"""